+++ src/trainers/trainers.service.ts
@@ -4,18 +4,19 @@ import { Trainer } from './entities/trainer.entity';
 import { Repository } from 'typeorm';
 import { CreateTrainerDto } from './dto/create-trainer.dto';
 import { UpdateTrainerDto } from './dto/update-trainer.dto';
+import { TrainerResponseDto } from './dto/trainer-response.dto';
 
 @Injectable()
 export class TrainersService {
     constructor(@InjectRepository(Trainer) private repository: Repository<Trainer>) {}
 
-    create(dto: CreateTrainerDto) {
+    async create(dto: CreateTrainerDto): Promise<TrainerResponseDto> {
         const trainer = this.repository.create(dto);
 
         return this.repository.save(trainer);
     }
 
-    async findAll() {
+    async findAll(): Promise<TrainerResponseDto[]> {
         // Busca todos os treinadores
         const trainers = await this.repository.find();
         
@@ -26,7 +27,7 @@ export class TrainersService {
         return trainers;
     }
 
-    async findOne(id: number) {
+    async findOne(id: number): Promise<TrainerResponseDto> {
         // Busca um treinador pelo id
         const trainer = await this.repository.findOne({ where: { id } });
         
@@ -37,7 +38,7 @@ export class TrainersService {
         return trainer;
     }
 
-    async update(id: number, dto: UpdateTrainerDto) {
+    async update(id: number, dto: UpdateTrainerDto): Promise<TrainerResponseDto> {
         // Busca um treinador pelo id
         const trainer = await this.findOne(id);
 
@@ -46,9 +47,13 @@ export class TrainersService {
         return this.repository.save(trainer);
     }
 
-    async remove(id: number) {
+    async remove(id: number): Promise<{ deleted: boolean }> {
         // Busca um treinador pelo id
-        const trainer = await this.findOne(id);
+        const trainer = await this.repository.findOne({ where: { id } });
+        
+        if (!trainer) {
+            throw new NotFoundException('Treinador não encontrado');
+        }
 
         await this.repository.remove(trainer);
